test vdo

+++ lognormal.py
@@ -262,7 +262,8 @@ def main():
 
     print('Creating BNAF model..')
     model = create_model(args, verbose=True)
-    model = model.double()
+    # model = model.double()
+
     print('Creating optimizer..')
     optimizer = torch.optim.Adam(model.parameters(), lr=args.learning_rate, amsgrad=True)
